Moved logic to ComponentInfoTrait
