+++ layers/GraphQLAPIForWP/plugins/graphql-api-for-wp/src/PluginSkeleton/PluginInfoTrait.php
@@ -4,27 +4,9 @@ declare(strict_types=1);
 
 namespace GraphQLAPI\GraphQLAPI\PluginSkeleton;
 
+use PoP\Root\Component\ComponentInfoTrait;
+
 trait PluginInfoTrait
 {
-    /** @var array<string, mixed> */
-    private static array $container = [];
-    private static bool $initialized = false;
-
-    /**
-     * @param array<string, mixed> $container
-     */
-    public static function init(array $container): void
-    {
-        if (self::$initialized) {
-            return;
-        }
-
-        self::$initialized = true;
-        self::$container = $container;
-    }
-
-    public static function get(string $key): mixed
-    {
-        return self::$container[$key] ?? null;
-    }
+    use ComponentInfoTrait;
 }
